+++ src/main/java/com/example/AplicatieMeditatii/customer/CustomerDao.java
@@ -5,11 +5,11 @@ import java.util.Optional;
 
 public interface CustomerDao {
     List<Customer> selectAllCustomers();
-    Optional<Customer> selectCustomerById(Integer id);
+    Optional<Customer> selectCustomerById(Long id);
     void insertCustomer(Customer customer);
     boolean existsPersonWithEmail(String email);
-    boolean existsPersonWithId(Integer id);
-    void deleteCustomerById(Integer id);
+    boolean existsPersonWithId(Long id);
+    void deleteCustomerById(Long id);
     void updateCustomer(Customer customer);
 
 }
