+++ src/main/java/com/example/AplicatieMeditatii/customer/CustomerRepository.java
@@ -3,7 +3,7 @@ package com.example.AplicatieMeditatii.customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CustomerRepository
-        extends JpaRepository<Customer, Integer> {
+        extends JpaRepository<Customer, Long> {
     boolean existsCustomerByEmail(String email);
-    boolean existsCustomerById(Integer id);
+    boolean existsCustomerById(Long id);
 }
